# project-code-main1/project-code-main/app/frontend/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "Country": {
        "description": "Represents a country with an ID, name, and flag.\n\nAttributes:\n-----------\n    id (str): The unique identifier for the country.\n    name (str): The name of the country.\n    flag (str): The URL of the flag image for the country.",
        "properties": {
          "flag": {
            "title": "Flag",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "flag"
        ],
        "title": "Country",
        "type": "object"
      },
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "title": "Error",
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "PlayerGroup33": {
        "description": "Represents a player in the system.\n\nAttributes:\n-----------\n- id (str): The unique identifier for the player.\n- name (str): The name of the player.\n- citizenships (Country): A countries that the player holds citizenship for.\n- DOB (str): The date of birth of the player.\n- height (str): The height of the player.\n- position (str): A position the player can play.\n- career_start (str): The year the player started their professional career.\n- image_url (str): A list of URL to images of the player.\n- teamId (str): The unique identifier for the team the player is currently playing for.\n- teamBadge_URL (str): The URL to the image of the badge of the team the player is currently playing for.\n- fbrefId (str): The id of the player on fbref",
        "properties": {
          "DOB": {
            "title": "Dob",
            "type": "string"
          },
          "career_start": {
            "title": "Career Start",
            "type": "string"
          },
          "citizenship": {
            "$ref": "#/components/schemas/Country"
          },
          "defending": {
            "title": "Defending",
            "type": "string"
          },
          "dribbling": {
            "title": "Dribbling",
            "type": "string"
          },
          "fbrefId": {
            "title": "Fbrefid",
            "type": "string"
          },
          "height": {
            "title": "Height",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "image_url": {
            "title": "Image Url",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "ovrall": {
            "title": "Ovrall",
            "type": "string"
          },
          "passing": {
            "title": "Passing",
            "type": "string"
          },
          "physicality": {
            "title": "Physicality",
            "type": "string"
          },
          "playmaking": {
            "title": "Playmaking",
            "type": "string"
          },
          "position": {
            "title": "Position",
            "type": "string"
          },
          "shooting": {
            "title": "Shooting",
            "type": "string"
          },
          "teamBadge_URL": {
            "title": "Teambadge Url",
            "type": "string"
          },
          "teamId": {
            "title": "Teamid",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "citizenship",
          "DOB",
          "height",
          "position",
          "career_start",
          "image_url",
          "teamId",
          "fbrefId",
          "teamBadge_URL",
          "ovrall",
          "dribbling",
          "shooting",
          "passing",
          "physicality",
          "playmaking",
          "defending"
        ],
        "title": "PlayerGroup33",
        "type": "object"
      },
...
      "PlayersResponseGroup33": {
        "description": "A response model for a list of players.\n\nAttributes:\n-----------\nplayerList (list[PlayerGroup33]): A list of Player objects.",
        "properties": {
          "playerList": {
            "items": {
              "$ref": "#/components/schemas/PlayerGroup33"
            },
            "title": "Playerlist",
            "type": "array"
          }
        },
        "required": [
          "playerList"
        ],
        "title": "PlayersResponseGroup33",
        "type": "object"
      },
...
      "ValidationErrorModel": {
        "properties": {
          "ctx": {
            "description": "an optional object which contains values required to render the error message.",
            "title": "Error context",
            "type": "object"
          },
          "loc": {
            "description": "the error's location as a list. ",
            "items": {
              "type": "string"
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "description": "a computer-readable identifier of the error type.",
            "title": "Message",
            "type": "string"
          },
          "type_": {
            "description": "a human readable explanation of the error.",
            "title": "Error Type",
            "type": "string"
          }
        },
        "title": "ValidationErrorModel",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Backend for the Sport Supergroup project",
    "title": "Sport Supergroup Backend",
    "version": "0.0.1"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/player/name/{name}": {
      "get": {
        "description": "Get players by name (e.g. Messi for all players named Messi)",
        "operationId": "get_players_by_Name_api_player_name__string_name__get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayersResponseGroup33"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ValidationErrorModel"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Retrieve players by Name.",
        "tags": [
          "players"
        ]
      }
    },
    "/api/player/{player_id}": {
      "get": {
        "description": "Get a player by player Q-number (e.g. Q50602 for a specific player)",
        "operationId": "get_player_by_ID_api_player__string_player_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "player_id",
            "required": true,
            "schema": {
              "title": "Player Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerGroup33"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ValidationErrorModel"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Retrieve a player by player Q-number.",
        "tags": [
          "players"
        ]
      }
    },
...
    "/api/players/{team_id}": {
      "get": {
        "description": "Get all players from a team by team Q-number (e.g. Q9439 for Arsenal F.C.) and their Q-numbers",
        "operationId": "get_team_players_api_players__string_team_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "team_id",
            "required": true,
            "schema": {
              "title": "Team Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayersResponseGroup33"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ValidationErrorModel"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Retrieve all players from a team by team Q-number and their Q-numbers.",
        "tags": [
          "players"
        ]
      }
    },
...
  },
...
}